+++ packages/theme_tailor_toolbox/lib/theme_tailor_toolbox.dart
@@ -1,0 +1,7 @@
+library theme_tailor_toolbox;
+
+export 'package:flutter/foundation.dart';
+export 'package:flutter/material.dart' show ThemeExtension;
+export 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
+
+export 'src/theme_tailor_encoder/encoder.dart';