+++ packages/theme_tailor/CHANGELOG.md
@@ -1,16 +1,18 @@
 # 1.0.4
 - Fix name collision when theme class contains `themes` field
+- Fix the bug where the generator could not handle encoders extending generic type encoders
+- Empty theme class will now generate with const constructor
 - Add option to generate constant themes
 
 # 1.0.3
 - Improve documentation,
-- Fix broken readme link
+- Fix the broken readme link
 - Add branding materials
 
 # 1.0.2
 - Improve documentation,
 - Add support for json_serializable,
-- Add support for nested themeExtension with `@themeExtension` annotation,
+- Add support for nested the extension with `@themeExtension` annotation,
 - Add support for Flutter's DiagnosticableTreeMixin (debugFillProperties)
 - Generated class will overwrite the `hashCode` and `==` operator. The equality operator will deeply compare collections in the class